centering text and correct windows bug with chrono

+++ game/include/network/packet_type.h
@@ -156,7 +156,7 @@ namespace game
 
     struct StartGamePacket : TypedPacket<PacketType::START_GAME>
     {
-        std::array<std::uint8_t, sizeof(unsigned long)> startTime{};
+        std::array<std::uint8_t, sizeof(unsigned long long)> startTime{};
     };
